Fixed the return type.

+++ Source/Engine/lobbysession.cpp
@@ -1,3 +1,3 @@
 #include "Engine/LobbySession.h"
 
-int(__cdecl* LobbySession::GetMaxClients)(unsigned int) = GetAddress<uint64(__cdecl*)(unsigned int localClientNumber /* This could be wrong... */)>(0x7FF7737E8CB0);
+int(__cdecl* LobbySession::GetMaxClients)(unsigned int) = GetAddress<int(__cdecl*)(unsigned int localClientNumber /* This could be wrong... */)>(0x7FF7737E8CB0);
